Add tool to convert curl timing to a .har file
This should be readable in Chrome, Firefox and websites like
http://www.softwareishard.com/har/viewer/.

Currently this is a very rough script, but should be enough for basic
testing. It assumes that all requests are made through the same
connection (not always true) and does not show the details of the
requests themselves (just `curl://request/1`, `curl://request/2`, ...).